Moved excavate into it's own folder
